Basic support for SIGINT handling
Previously, if one interrupted an executing command line, it would
simply exit. This meant, things that we might want to clean-up, and
allow for a more graceful shut down weren't possible.

Now, there is a signal handling code, which executes prior to shut down.
Tie this in wasn't easy as it requires the handler to exit the program
within the handler, and previously the whole command line program was
written to execute and return an exit code. The whole exit process code
had to be brought in as an injectable effect function, so we can
override it during test.

A remaining issue is that the code could do with some clean-up, there is
a fair bit of indentation, which is usuall indicative of code with lots
of conditionals, and complex logic. Further more, the code coverage is
low, a key area that's missing is the signal handling logic, which is
troubling as bugs hide in such important, but rarely exercised code
paths.