First bit of work to get daemon support
We can fire up command lines applications with short lifetimes, but
running a daemon is a slightly different matter. Likely, this will
necessitate a number of changes to the API to make it more pleasant to
use, and figure out how to test a running daemon that gets interrupted
and shutsdown, and how to test a daemon gracefully shutting down on its
own.