Tests for Console work again
The issue right now is that there is a bug in Clikt 1.5.0 where the
Context.console, which controls where output goes, is not passed down
to subCommands from the parent so it's not possible to test subCommand
output.

Although there might be another way with a testing trick that Clikt
itself uses to capture output.